requirements.txt
github3.py>=0.7.0
PyYaml
